+++ ibm_spectrum_lsf/datadog_checks/ibm_spectrum_lsf/check.py
@@ -7,6 +7,8 @@ from datadog_checks.base import AgentCheck
 from .client import LSFClient
 from .config_models import ConfigMixin
 from .processors import (
+    BadminPerfmonProcessor,
+    BHistProcessor,
     BHostsProcessor,
     BJobsProcessor,
     BQueuesProcessor,
@@ -34,17 +36,19 @@ class IbmSpectrumLsfCheck(AgentCheck, ConfigMixin):
     def parse_config(self):
         self.tags = list[str](self.config.tags or []) + [f"lsf_cluster_name:{self.config.cluster_name}"]
 
-    def initialize_processors(self):
+    def initialize_processors(self) -> None:
         self.processors = [
-            LsClustersProcessor(self.client, self.log, self.tags),
-            LSHostsProcessor(self.client, self.log, self.tags),
-            LsLoadProcessor(self.client, self.log, self.tags),
-            BHostsProcessor(self.client, self.log, self.tags),
-            BJobsProcessor(self.client, self.log, self.tags),
-            BQueuesProcessor(self.client, self.log, self.tags),
-            BSlotsProcessor(self.client, self.log, self.tags),
-            GPULoadProcessor(self.client, self.log, self.tags),
-            GPUHostsProcessor(self.client, self.log, self.tags),
+            LsClustersProcessor(self.client, self.config, self.log, self.tags),
+            LSHostsProcessor(self.client, self.config, self.log, self.tags),
+            LsLoadProcessor(self.client, self.config, self.log, self.tags),
+            BHostsProcessor(self.client, self.config, self.log, self.tags),
+            BJobsProcessor(self.client, self.config, self.log, self.tags),
+            BQueuesProcessor(self.client, self.config, self.log, self.tags),
+            BSlotsProcessor(self.client, self.config, self.log, self.tags),
+            BHistProcessor(self.client, self.config, self.log, self.tags),
+            GPULoadProcessor(self.client, self.config, self.log, self.tags),
+            GPUHostsProcessor(self.client, self.config, self.log, self.tags),
+            BadminPerfmonProcessor(self.client, self.config, self.log, self.tags),
         ]
 
     def check(self, instance):
@@ -57,6 +61,17 @@ class IbmSpectrumLsfCheck(AgentCheck, ConfigMixin):
             return
 
         for processor in self.processors:
-            metrics = processor.process_metrics()
-            for metric in metrics:
-                self.gauge(metric.name, metric.value, tags=metric.tags)
+            if processor.should_run():
+                metrics = processor.process_metrics()
+                for metric in metrics:
+                    self.gauge(metric.name, metric.value, tags=metric.tags)
+            else:
+                self.log.trace("Skipping %s metrics; excluded in configuration.", processor.name)
+
+    def cancel(self):
+        if (
+            self.config.metric_sources
+            and 'badmin_perfmon' in self.config.metric_sources
+            and self.config.badmin_perfmon_auto
+        ):
+            self.client.badmin_perfmon_stop()