+++ ibm_spectrum_lsf/datadog_checks/ibm_spectrum_lsf/config_models/instance.py
@@ -34,10 +34,12 @@ class InstanceConfig(BaseModel):
         arbitrary_types_allowed=True,
         frozen=True,
     )
+    badmin_perfmon_auto: Optional[bool] = None
     cluster_name: str
     disable_generic_tags: Optional[bool] = None
     empty_default_hostname: Optional[bool] = None
     metric_patterns: Optional[MetricPatterns] = None
+    metric_sources: Optional[tuple[str, ...]] = None
     min_collection_interval: Optional[float] = None
     service: Optional[str] = None
     tags: Optional[tuple[str, ...]] = None