+++ karpenter/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 <!-- towncrier release notes start -->
 
+## 3.2.0 / 2025-11-26
+
+***Added***:
+
+* Add support for Karpenter v1.8 metrics ([#21819](https://github.com/DataDog/integrations-core/pull/21819))
+* Bump minimum version of datadog-checks-base to 37.24.0 ([#21945](https://github.com/DataDog/integrations-core/pull/21945))
+
 ## 3.1.1 / 2025-10-31
 
 ***Fixed***: